knowledge_graph/requirements.txt
neo4j>=5.0.0
langchain-neo4j>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.0.0
openai>=1.0.0
